==> notebooks/03_transform_ts_data_into_features_and_targets.ipynb <==
## Here we transform the time series data into tabular data by using sliding window technique

import pandas as pd
import numpy as np
from tqdm import tqdm

ts_data = pd.read_parquet("../data/transformed/ts_data_2022_01.parquet")

ts_data

ts_data_one_location = ts_data.loc[ts_data["pickup_location_id"] == 43].reset_index(drop=True)

ts_data_one_location.head(15)

def get_cutoff_indices(data: pd.DataFrame, n_features: int, step_size:int) -> list:
    stop_position = len(data) - 1
    
    # start the first sub-sequence at index position 0
    subseq_first_idx = 0
    subseq_mid_idx = n_features
    subseq_last_idx = n_features + 1
    indices_lst = []
    
    while subseq_last_idx <= stop_position:
        indices_lst.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))
        
        subseq_first_idx += step_size
        subseq_mid_idx += step_size
        subseq_last_idx += step_size
        
    return indices_lst

n_features = 24
step_size = 1

indices_lst = get_cutoff_indices(ts_data_one_location, n_features, step_size)

indices_lst[:5]

n_examples = len(indices_lst)
# features
x = np.ndarray(shape = (n_examples, n_features), dtype = np.float32)
# target
y = np.ndarray(shape = (n_examples, 1), dtype = np.float32)
pickup_hours_lst = []

for i, idx in enumerate(indices_lst):
    x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]["rides_count"].values
    y[i] = ts_data_one_location.iloc[idx[1]:idx[2]]["rides_count"].values
    pickup_hours_lst.append(ts_data_one_location.iloc[idx[1]]["pickup_hour"])

print(f"{x.shape = }")
print(f"{x = }")
print(f"{pickup_hours_lst[:5] = }")

Transform numpy array to DataFrame

features_one_location = pd.DataFrame(x, columns = [f"rides_previous_{i + 1}_hour" for i in reversed(range(n_features))])

features_one_location

Tranform target "y" numpy array into a pandas dataframe

targets_one_location = pd.DataFrame(y, columns = ["target_rides_next_hour"])

targets_one_location

Create a funtion that does the data procesing for all locations. this funtion will tranform time series data into tabular data

def tranform_ts_data_into_features_and_target(ts_data: pd.DataFrame, input_seq_len: int, step_size: int) -> pd.DataFrame:
    """
    Slices and transposes data from time-series format into a (features, target) format
    that can use to train supervised ML models
    """
    assert set(ts_data.columns) == {"pickup_hour", "rides_count", "pickup_location_id"}
    
    locations_ids = ts_data["pickup_location_id"].unique()
    features = pd.DataFrame()
    target = pd.DataFrame()
    
    
    for location_id in tqdm(locations_ids):
        
        # keep only ts data for this location_id
        ts_data_one_location = ts_data.loc[ts_data["pickup_location_id"] == location_id, ["pickup_hour", "rides_count"]]
        
        # pre-compute cutoff indices to split datafrane rows
        indices = get_cutoff_indices(ts_data_one_location, input_seq_len, step_size)
        
        # slice and transpose data into numpy arrays for features and target
        n_examples = len(indices)
        x = np.ndarray(shape = (n_examples, input_seq_len), dtype = np.float32)
        y = np.ndarray(shape = (n_examples), dtype = np.float32)
        pickup_hours_lst = []
        
        for i, idx in enumerate(indices):
            x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]["rides_count"].values
            y[i] = ts_data_one_location.iloc[idx[1]:idx[2]]["rides_count"].values
            pickup_hours_lst.append(ts_data_one_location.iloc[idx[1]]["pickup_hour"])
            
        # convert numpy arrays into pandas dataframes for both features and target
        features_one_location = pd.DataFrame(x, columns = [f"rides_previous_{i + 1}_hour" for i in reversed(range(input_seq_len))])
        features_one_location["pickup_hour"] = pickup_hours_lst
        features_one_location["pickup_location_id"] = location_id
        
        # target
        targets_one_location = pd.DataFrame(y, columns = [f"target_rides_next_hour"])
        
        # concatenate features and target into 2 dataframes
        features = pd.concat([features, features_one_location])
        target = pd.concat([target, targets_one_location])
        
    features.reset_index(drop=True, inplace=True)
    target.reset_index(drop=True, inplace=True)
    
    return features, target["target_rides_next_hour"]

# call the function

# 24 hours, 7 days, 1 week ==> 1 week of history
features_df, target_df = tranform_ts_data_into_features_and_target(ts_data, input_seq_len = 24*7*1, step_size = 24)

print(f"{features_df.shape = }")
print(f"{target_df.shape = }")
