==> notebooks/04_tranform_row_data_into_features_and_target.ipynb <==
#from src import *

import os

os.chdir('d:/Courses/Pau_ML_course/Taxi_demand_predictor')

os.getcwd()

# reload everytime we import a new module in .py file
%reload_ext autoreload
%autoreload 2

from src.data_processing import load_raw_data, transform_raw_data_into_ts_data



### Fetch all the data of year 2022

rides = load_raw_data(year= 2022)

rides

ts_data = transform_raw_data_into_ts_data(rides)

ts_data

### Transform the Time series data into tabular data 

from src.data_processing import tranform_ts_data_into_features_and_target

features, target = tranform_ts_data_into_features_and_target(
    ts_data, 
    input_seq_len= 24*28*1, # one month
    step_size= 24
)


print(f"{features.shape = }")
print(f"{target.shape = }")

tabular_data = features
tabular_data["target_rides_next_hour"] = target

from src.paths import TRANSFORMED_DATA_DIR

tabular_data.to_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
