==> notebooks/06_baseline_model.ipynb <==
import os

os.chdir('../')

%pwd

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

### Split into train and test data

from datetime import datetime
from src.data_split import train_test_split

Cuttoff date 1st June 2022

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date = datetime(2022, 6, 1, 0, 0, 0),
    target_column_name = "target_rides_next_hour"
)


print(f"X_train shape: {X_train.shape}")
print(f"y_train shape: {y_train.shape}")
print(f"X_test shape: {X_test.shape}")
print(f"y_test shape: {y_test.shape}")

X_train.head()

### Baseline model 1

import numpy as np

class BaseLineModelPreviousHour:
    """Prediction = actual demand observed in the last hour
    """
    
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        """no training required for base line model

        Args:
            X_train (pd.DataFrame): _description_
            y_train (pd.Series): _description_
        """
        pass
    
    
    def predict(self, X_test: pd.DataFrame) -> np.ndarray:
        """Predict the demand for the next hour based on the actual demand observed in the last hour.

        Args:
            X_test (pd.DataFrame): _description_

        Returns:
            np.ndarray: _description_
        """
        
        return X_test[f"rides_previous_1_hour"]

### Create object of the base line class

base_model = BaseLineModelPreviousHour()
predictions = base_model.predict(X_test)
predictions

X_test["rides_previous_1_hour"]

### Error Metric: MAE

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f"test MAE: {test_mae:.4f}")

### Baseline Model 2 : observe the demand from last week

class BaselinModelPreviousWeek:
    """Prediction = actual demand observed at t - 7 days (observe the last week data to predict)
    """
    
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
    
    
    def predict(self, X_test: pd.DataFrame) -> np.ndarray:
        return X_test[f"rides_previous_{7 * 24}_hour"]

model_2 = BaselinModelPreviousWeek()
predictions = model_2.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)
print(f"Baseline model 2---> test MAE: {test_mae:.4f}")

2nd baseline model is an improvement over the 1st baseline

### Baseline Model 3 - take avg of the last 4 weeks

class BaselineModelLast4Weeks:
    """ 
    Prediction = actual deman observed at t - 7 days, t - 14 days, t - 21 days, t - 28 days
    """
    
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
    
    
    def predict(self, X_test: pd.DataFrame) -> pd.Series:
        
        """take avg of the last 4 weeks demand to predict the next hour demand
        """
        return 0.25 * (
            X_test[f"rides_previous_{7 * 24}_hour"] + \
            X_test[f"rides_previous_{2 * 7 * 24}_hour"] + \
            X_test[f"rides_previous_{3 * 7 * 24}_hour"] + \
            X_test[f"rides_previous_{4 * 7 * 24}_hour"]
        )
    
    

model_3 = BaselineModelLast4Weeks()
predictions = model_3.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)
print(f"Baseline model 3---> test MAE: {test_mae:.4f}")

So far, baseline model 3 gave the best MAE score. So, now we can compare the ML models with this baseline score
